update run.sh to pass argument to the docker build or docker-compose build commands. (Need to pass '--no-cache' so the latest changes were applied) Add Playwright and Protractor tutorials

+++ packages/eyes-playwright/tutorial/run.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 set -e
-rm -r ./package
+[ -d "./package" ] && rm -r ./package
 mkdir package
 cd ..
 yarn pack
@@ -8,5 +8,5 @@ package=$(find applitools*.tgz)
 mv "$package" ./tutorial/package/"$package"
 cd ./tutorial
 docker-compose build $1
-docker-compose run wdio5_selenium_basic
-docker-compose run wdio5_selenium_ultrafastgrid
+docker-compose run playwright_basic
+docker-compose run playwright_ultrafastgrid
